Rename platform/platform to index removing default
