--- src/platform/platforms.js
@@ -1,11 +1,0 @@
-import BasePlatform from './platform.base';
-import BasicPlatform from './platform.basic';
-import DomPlatform from './platform.dom';
-
-export {BasicPlatform, DomPlatform, BasePlatform};
-
-/**
- * @namespace Chart.platforms
- * @see https://chartjs.gitbooks.io/proposals/content/Platform.html
-*/
-export default {BasicPlatform, DomPlatform, BasePlatform};
